// apps/web/src/gen/types/DailyTotal.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type DailyTotal = {
  /**
   * @type number
   */
  mdr: number
  /**
   * @type number
   */
  tpv: number
  /**
   * @type number
   */
  rav: number
  /**
   * @type number
   */
  liquido: number
  /**
   * @type string
   */
  data_recebimento: string
}